tests/requirements.txt
pytest==7.4.0
pytest-html==3.2.0
pytest-json-report==1.5.0
requests==2.31.0
python-keycloak==3.7.0
PyJWT==2.8.0
cryptography==41.0.3
lxml==4.9.3
xmlsec==1.3.13
python-saml==1.15.0
selenium==4.11.2
webdriver-manager==4.0.0